requirements.txt
transformers
requests
python-dotenv